Initial commit
`hips` is a tool meant to enabling storing of various production
secretw alongside code in a git repository. We cannot track those
in plain text because code repositories are not well protected.

The solution is to store them anyway, but to encrypt them in order to
prevent anyone without the master password from seeing any of the
production secrets.

This tool helps by curating a list of key/value pairs with the values
encrypted seamlessly. Once this is done, everything can be stored in
the repository except for the master password.

The tool is still lacking the encryption capabilities.